+++ dynchk/libs/libc/readv.c
@@ -4,9 +4,9 @@
 #include <dlfcn.h>
 #include <stddef.h>
 #undef readv
-static int(*funcptr) (int , const struct iovec * , size_t ) = 0;
+static ssize_t(*funcptr) (int , const struct iovec * , int ) = 0;
 
-int readv (int arg0 , const struct iovec * arg1 , size_t arg2 )
+ssize_t readv (int arg0 , const struct iovec * arg1 , int arg2 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "readv");
@@ -16,7 +16,7 @@ int readv (int arg0 , const struct iovec * arg1 , size_t arg2 )
 	return funcptr(arg0, arg1, arg2);
 }
 
-int lsb_readv (int arg0 , const struct iovec * arg1 , size_t arg2 )
+ssize_t lsb_readv (int arg0 , const struct iovec * arg1 , int arg2 )
 {
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "readv");
